refactor: actions -> serverActions
